Fix hardcoding height and bottom values for dungeons
Worlds whose generation doesn't have a height of 256 and a bottom of 0
will not work with dungeon seed cracking. This is because the dungeon
seed cracking in this mod always assumes that the generation has a
height of 256 and a bottom of 0.

To fix this, just use getHeight() instead of 256, and subtract
getBottomY() from the Y-position of the dungeon. We need to do this,
because to generate the Y-position of the dungeon, Minecraft will do
Random.nextInt(getHeight()) and then add getBottomY() to the result.
(You cannot specify the minimum bound with Random.nextInt(), and the
minimum bound will always default to 0; so this is the next best thing
to do.) And the reverse of adding is subtracting.

Also, to be able to use getHeight() and getBottomY(), we'll need to pass
the World around a bit, but this is no big deal.

This makes the seed cracker mod work with the Caves & Cliffs 1.18 world
generation preview data pack (which generates worlds with a height of
384 and a bottom of -64), completely flawlessly. I tested it.